Refactored code to remove code in headers
