+++ sdk/attestation/azure-security-attestation/CMakeLists.txt
@@ -33,22 +33,25 @@ set(
     inc/azure/attestation.hpp
     inc/azure/attestation/attestation_client_models.hpp
     inc/azure/attestation/attestation_client_options.hpp
+    
     src/private/package_version.hpp
     src/private/attestation_common_request.hpp
     src/private/attestation_deserializer.hpp
-    src/private/jsonwebkeyset.hpp 
+     
     src/private/attestation_client_models_private.hpp
     src/private/crypto/inc/crypto.hpp
     src/private/crypto/inc/cryptocert.hpp
     src/private/crypto/inc/cryptokeys.hpp
- )
+    src/private/attestation_client_private.hpp)
 
 set(
   AZURE_ATTESTATION_SOURCE
     src/attestation_client.cpp
     src/attestation_client_options.cpp
     src/attestation_common_request.cpp
-)
+    src/private/attestation_deserializers.cpp
+    src/private/jsonhelpers.cpp
+    src/private/attestation_client_private.cpp)
 
 #if(WIN32)
 #set(
@@ -68,7 +71,7 @@ set(
 
 create_per_service_target_build(azure-security-attestation attestation)
 
-add_library(azure-security-attestation ${AZURE_ATTESTATION_HEADER} ${AZURE_ATTESTATION_SOURCE})
+add_library(azure-security-attestation ${AZURE_ATTESTATION_HEADER} ${AZURE_ATTESTATION_SOURCE} "src/private/attestation_client_private.hpp")
 
 #if(WIN32)
 #    find_package(WIL CONFIG REQUIRED)
