+++ sdk/attestation/azure-security-attestation/src/attestation_client.cpp
@@ -8,6 +8,7 @@
 #include "azure/core/http/policies/policy.hpp"
 #include "azure/core/internal/http/pipeline.hpp"
 #include "private/attestation_client_models_private.hpp"
+#include "private/attestation_client_private.hpp"
 #include "private/attestation_common_request.hpp"
 #include "private/attestation_deserializer.hpp"
 #include "private/package_version.hpp"
